blah

+++ image_set/generate_dataset.py
@@ -18,7 +18,7 @@ for n,i in enumerate(imgs):
     k = mmreadgray(i)
     size = k.shape
     scale = 4000.0/min(size)
-    newsize = (size[0]*scale, size[1]*scale)
+    newsize = (int(size[0]*scale), int(size[1]*scale))
 
     img = (mmhmin(k, 50) > 128).astype(numpy.uint8)*255
     img = mmareaclose(img, 20)
@@ -26,7 +26,7 @@ for n,i in enumerate(imgs):
     pil_img = Image.fromarray(img, "L")
     pil_img = pil_img.resize(newsize)
     pil_img.save( "output/"+f+".png" )
-    os.system("convert -compress none %s %s" % 
+    os.system("convert -compress none %s %s" %
               ("output/"+f+".png", "output/"+f+".pgm"))
 
 import sys
